+++ jlox/include/lox.h
@@ -14,7 +14,7 @@ public:
 
 	static void Error(const Token& token, const std::string& message);
 
-	static void Error(int line, const std::string& message);
+	static void Error(size_t line, const std::string& message);
 
 	static void runtimeError(const RuntimeError& error);
 
@@ -28,7 +28,7 @@ private:
 
 	static void Run(const std::string& source);
 
-	static void Report(int line, const std::string& where, const std::string& message);
+	static void Report(size_t line, const std::string& where, const std::string& message);
 
 	static void PrintStatus();
 };
