+++ jlox/include/loxCallable.h
@@ -11,5 +11,5 @@ public:
 	virtual ~LoxCallable() = default;
 
 	virtual object_t call(Interpreter* interpreter, std::vector<object_t> arguments) const = 0;
-	virtual int arity() const = 0;
+	virtual size_t arity() const = 0;
 };
